Factor out binding command execution to separate file
